gui improvements

+++ src/main/java/it/polimi/ingsw/sagrada/gui/utils/GUIManager.java
@@ -133,7 +133,7 @@ public class GUIManager {
         return perc * getGameWindowHeight() / 100;
     }
 
-    public static double getGameWidthPixel(int perc) {
+    public static double getGameWidthPixel(double perc) {
         return perc * getGameWindowWidth() / 100;
     }
